requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
pytest>=7.0.0
pytest-cov>=4.0.0
